feat: implement arm64 JIT compilation for I32_CONST and RETURN instructions
