+++ asm/arm64/reg.go
@@ -32,39 +32,6 @@ const (
 	X28 = 28
 	X29 = 29
 	X30 = 30
-
-	W0  = 0
-	W1  = 1
-	W2  = 2
-	W3  = 3
-	W4  = 4
-	W5  = 5
-	W6  = 6
-	W7  = 7
-	W8  = 8
-	W9  = 9
-	W10 = 10
-	W11 = 11
-	W12 = 12
-	W13 = 13
-	W14 = 14
-	W15 = 15
-	W16 = 16
-	W17 = 17
-	W18 = 18
-	W19 = 19
-	W20 = 20
-	W21 = 21
-	W22 = 22
-	W23 = 23
-	W24 = 24
-	W25 = 25
-	W26 = 26
-	W27 = 27
-	W28 = 28
-	W29 = 29
-	W30 = 30
-
-	SP = 31
-	LR = 30 // X30과 동일
+	SP  = 31
+	LR  = 30 // X30과 동일
 )
